Changes after CR

+++ src/main/java/Objects/Square.java
@@ -5,7 +5,7 @@ public class Square implements SurfaceArea {
     double a;
     private double surfaceArea;
 
-    Square(float a) {
+    Square(double a) {
         this.a = a;
         this.surfaceArea = calculateSurfaceArea();
     }
